+++ include/brdf/microfacet_brdf.h
@@ -10,9 +10,9 @@
 
 class MicrofacetBRDF final : public BRDF {
 public:
-    [[nodiscard]] float eval(const Polar &wi, const Polar &wo) const override;
+    [[nodiscard]] double eval(const Polar &wi, const Polar &wo) const override;
 
-    [[nodiscard]] float pdf(const Polar &wi, const Polar &wo) const override;
+    [[nodiscard]] double pdf(const Polar &wi, const Polar &wo) const override;
 
     [[nodiscard]] Polar sample(const Polar &wi) const override;
