+++ include/utils.h
@@ -12,7 +12,7 @@
 #include <stdio.h>
 
 namespace utils {
-    float safeSqrt(float x);
+    double safeSqrt(double x);
 
     template<typename T>
     T mse(const T &mean, const std::vector<T> &samples) {
@@ -22,7 +22,7 @@ namespace utils {
             // std::cout << "Sample: " << sample << std::endl;
             sum += (sample - mean) * (sample - mean);
         }
-        return sum / static_cast<float>(samples.size());
+        return sum / static_cast<double>(samples.size());
     }
 
     /// Vec2f
@@ -30,43 +30,43 @@ namespace utils {
 
     Vec2f reflect(const Vec2f &v, const Vec2f &axis);
 
-    float dot(const Vec2f &v1, const Vec2f &v2);
+    double dot(const Vec2f &v1, const Vec2f &v2);
 
     Vec2f normalize(const Vec2f &v);
 
-    float cosTheta(const Vec2f &v);
+    double cosTheta(const Vec2f &v);
 
-    float sinTheta(const Vec2f &v);
+    double sinTheta(const Vec2f &v);
 
-    float tanTheta(const Vec2f &v);
+    double tanTheta(const Vec2f &v);
 
     /// Vec3f
     Spherical toSpherical(const Vec3f &);
 
     Vec3f reflect(const Vec3f &v, const Vec3f &axis);
 
-    float dot(const Vec3f &v1, const Vec3f &v2);
+    double dot(const Vec3f &v1, const Vec3f &v2);
 
     Vec3f normalize(const Vec3f &v);
 
-    float cosTheta(const Vec3f &v);
+    double cosTheta(const Vec3f &v);
 
-    float sinTheta(const Vec3f &v);
+    double sinTheta(const Vec3f &v);
 
-    float cosPhi(const Vec3f &v);
+    double cosPhi(const Vec3f &v);
 
-    float sinPhi(const Vec3f &v);
+    double sinPhi(const Vec3f &v);
 
     /// Polar
     Vec2f toVec(const Polar &);
 
     Polar reflect(const Polar &p, const Polar &axis);
 
-    float cosTheta(const Polar &p);
+    double cosTheta(const Polar &p);
 
-    float sinTheta(const Polar &p);
+    double sinTheta(const Polar &p);
 
-    float tanTheta(const Polar &p);
+    double tanTheta(const Polar &p);
 
     Polar normalize(const Polar &p);
 
@@ -75,7 +75,7 @@ namespace utils {
 
     Vec3f hemisphereSample();
 
-    inline float lerp(const float v0, const float v1, const float t) {
+    inline double lerp(const double v0, const double v1, const double t) {
         return v0 + t * (v1 - v0);
     }
 }
