+++ source/renderers/mis.cpp
@@ -5,32 +5,32 @@
 #include "renderers/mis.h"
 
 // TODO cosTheta wi
-float sampling::mis(std::uint32_t iterations, const MSTree &msTree, const BRDF &brdf,
+double sampling::mis(std::uint32_t iterations, const MSTree &msTree, const BRDF &brdf,
                     const std::vector<LightSource> &lightSources, const Polar &wi) {
-    float color = 0.f;
+    double color = 0.0;
     for (int i = 0; i < iterations; ++i) {
         color += mis(msTree, brdf, lightSources, wi);
     }
-    color /= static_cast<float>(iterations);
+    color /= static_cast<double>(iterations);
     return color;
 }
 
-float mix(float val0, float w0, float val1, float w1) {
+double mix(double val0, double w0, double val1, double w1) {
     return (val0 * w0 + val1 * w1) / (w0 + w1);
 }
 
 // TODO cosTheta wo
-float sampling::mis(const MSTree &msTree, const BRDF &brdf, const std::vector<LightSource> &lightSources,
+double sampling::mis(const MSTree &msTree, const BRDF &brdf, const std::vector<LightSource> &lightSources,
                     const Polar &wi) {
     const Polar wo = brdf.sample(wi);
-    const float brdfIncoming = intersect_lights(lightSources, wo);
-    const float brdfPdf = brdf.pdf(wi, wo);
-    const float brdfColor = brdfIncoming * brdf.eval(wi, wo) / brdfPdf;
+    const double brdfIncoming = intersect_lights(lightSources, wo);
+    const double brdfPdf = brdf.pdf(wi, wo);
+    const double brdfColor = brdfIncoming * brdf.eval(wi, wo) / brdfPdf;
 
     const MSTSample sample = msTree.sample();
-    const float mstIncoming = intersect_lights(lightSources, sample.wo);
-    const float mstPdf = sample.pdf;
-    const float mstColor = mstIncoming * brdf.eval(wi, sample.wo) / mstPdf;
+    const double mstIncoming = intersect_lights(lightSources, sample.wo);
+    const double mstPdf = sample.pdf;
+    const double mstColor = mstIncoming * brdf.eval(wi, sample.wo) / mstPdf;
 
     // (
     //   (color0 / pdf0) * pdf0 +
